Removed tests from landing page to separate component with route
