+++ client/src/App.js
@@ -9,6 +9,7 @@ import TablesPage from './components/TablesPage';
 import NotFound from './components/NotFound';
 import RequireNotAuth from './components/HOC/RequireNotAuth';
 import RequireAuth from './components/HOC/RequireAuth';
+import Test from './components/Test';
 
 class App extends Component {
   render() {
@@ -20,6 +21,7 @@ class App extends Component {
           <Route path="/register" component={RequireNotAuth(Register)} />
           <Route path="/tables" component={RequireAuth(TablesPage)} />
           <Route path="/servers" component={RequireAuth(Servers)} />
+          <Route path="/test" component={RequireAuth(Test)} />
           <Route path="/404" component={NotFound} exact />
           <Redirect to="/404" />
         </Switch>
